// src/DockiUp.Frontend/src/app/api/model/projectSettingsDto.ts
/**
 * DockiUp.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectSecretDto } from './projectSecretDto';


export interface ProjectSettingsDto { 
    isGit: boolean;
    gitCredentialId: string | null;
    secrets: Array<ProjectSecretDto>;
}